Disable gitlab endpoints if not needed
